moved all human moves out of game class
