+++ src/playerOptions/computer.h
@@ -7,7 +7,7 @@ class Computer : public PlayerType{
 
 public:
     Computer(std::string colour);
-    virtual void playerMove();
+    virtual bool playerMove(std::string input, Board &gameBoard);
 
 };
