+++ src/playerOptions/computerL1.cc
@@ -1,12 +1,13 @@
 #include "computerL1.h"
+#include "../board/board.h"
 
-ComputerL1::ComputerL1(std::string colour) : Computer{colour} {
+// ComputerL1::ComputerL1(std::string colour) : Computer{colour} {
 
-}
-
-
+// }
 
 
-void ComputerL1::playerMove() {
 
-}
+
+// bool ComputerL1::playerMove(std::string input, std::string colour, Board &gameBoard) {
+//     return true;
+// }
