+++ src/playerOptions/computerL3.h
@@ -7,7 +7,7 @@ class ComputerL3 : public Computer {
 
 public:
     ComputerL3(std::string colour);
-    void playerMove() override;
+    bool playerMove(std::string input, Board &gameBoard) override;
 
 };
