+++ src/playerOptions/playerType.h
@@ -2,16 +2,23 @@
 #define PLAYERTYPE_H
 
 #include <string>
+#include <vector>
+#include <memory>
+
+#include "../pieces/piece.h"
+
+class Board;
 
 class PlayerType {
 
-private:
+protected:
     std::string colour;
+
 
 public:
 
     PlayerType(std::string colour);
-    virtual void playerMove();
+    virtual bool playerMove(std::string input, Board &gameBoard);
     virtual ~PlayerType() = 0;
     std::string getColour();
